增加转码DFX

+++ services/engine/histreamer/transcoder/hitranscoder_impl.cpp
@@ -312,7 +312,7 @@ Status HiTransCoderImpl::ConfigureInputVideoMetaData(const std::vector<std::shar
     const size_t &index)
 {
     MEDIA_LOG_I("InputVideo contains videoTrack");
-    FALSE_RETURN_V_MSG_E(trackInfos.size() > 0, static_cast<int32_t>(Status::ERROR_INVALID_PARAMETER),
+    FALSE_RETURN_V_MSG_E(trackInfos.size() > 0, Status::ERROR_INVALID_PARAMETER,
         "trackInfos are invalid.");
     isExistVideoTrack_ = true;
     Plugins::VideoRotation rotation = Plugins::VideoRotation::VIDEO_ROTATION_0;
@@ -350,7 +350,7 @@ Status HiTransCoderImpl::ConfigureInputVideoMetaData(const std::vector<std::shar
 
 Status HiTransCoderImpl::ConfigureMetaData(const std::vector<std::shared_ptr<Meta>> &trackInfos)
 {
-    FALSE_RETURN_V_MSG_E(trackInfos.size() > 0, static_cast<int32_t>(Status::ERROR_INVALID_PARAMETER),
+    FALSE_RETURN_V_MSG_E(trackInfos.size() > 0, Status::ERROR_INVALID_PARAMETER,
         "trackInfos are invalid.");
     for (size_t index = 0; index < trackInfos.size(); index++) {
         std::shared_ptr<Meta> meta = trackInfos[index];
